# sample/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample REST API Server.",
        "title": "GAIA Sample API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
        "/department": {
            "get": {
                "description": "Get all departments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Get all departments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "departments": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/v1.departmentRespDto"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create new department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Create new department",
                "parameters": [
                    {
                        "description": "Department data",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.departmentReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "department": {
                                                            "$ref": "#/definitions/v1.departmentRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/department/{deptId}": {
            "get": {
                "description": "Get department by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Get department by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "department": {
                                                            "$ref": "#/definitions/v1.departmentRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update department by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Update department by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Department data",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.departmentReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "department": {
                                                            "$ref": "#/definitions/v1.departmentRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete department by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Delete department by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "department": {
                                                            "$ref": "#/definitions/v1.departmentRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/department/{deptId}/employee": {
            "get": {
                "description": "Get all employees",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Get all employees",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "employees": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/v1.employeeRespDto"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create new employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Create new employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "employee": {
                                                            "$ref": "#/definitions/v1.employeeRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/department/{deptId}/employee/{employeeId}": {
            "get": {
                "description": "Get employee by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Get employee by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "employee": {
                                                            "$ref": "#/definitions/v1.employeeRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update employee by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Update employee by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Employee data",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.employeeReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "employee": {
                                                            "$ref": "#/definitions/v1.employeeRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete employee by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Delete employee by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.ResponseDoc"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/v1.DataDoc"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "employee": {
                                                            "$ref": "#/definitions/v1.employeeRespDto"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.DataDoc": {
            "type": "object"
        },
        "v1.ResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.DataDoc"
                },
                "message": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "v1.departmentReqDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "v1.departmentRespDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "v1.employeeReqDto": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string"
                },
                "employeeNumber": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "v1.employeeRespDto": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "employeeNumber": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                }
            }
        }
    }
}